+++ config_example.yml
@@ -1,0 +1,4 @@
+# Discord access token
+# https://discordapp.com/developers/applications/
+BOT_TOKEN:
+
