+++ requirements.txt
@@ -1,7 +1,16 @@
 aiohttp==3.4.4
+async-timeout==3.0.1
+attrs==18.2.0
 beautifulsoup4==4.6.3
-discord.py==1.0.0a1580+g836ae73
+chardet==3.0.4
+discord.py==1.0.0a1619+g1f2602d
+idna==2.7
+idna-ssl==1.1.0
 lxml==4.2.5
+multidict==4.5.1
+pkg-resources==0.0.0
 py-stackexchange==2.2.7
-PyNaCl==1.2.1
 PyYAML==3.13
+six==1.11.0
+websockets==6.0
+yarl==1.2.6
